fix: warning about __main__ already being imported
